Create a helper class for BlurConfig that provides min and max blur radius
This changes all constants used in the bouncer transition to use the same values that are used by BlurUtils,
we were using a hardcoded value that resulted in the blur on shade and bouncer being slightly different.

Other changes:
 - Fixes bouncer -> dozing transition flicker

Bug: 370555003
Test: Updated unit tests
Flag: com.android.systemui.bouncer_ui_revamp

+++ packages/SystemUI/multivalentTests/src/com/android/systemui/statusbar/NotificationShadeDepthControllerTest.kt
@@ -111,8 +111,8 @@ class NotificationShadeDepthControllerTest : SysuiTestCase() {
             answer.arguments[0] as Float / maxBlur.toFloat()
         }
         `when`(blurUtils.supportsBlursOnWindows()).thenReturn(true)
-        `when`(blurUtils.maxBlurRadius).thenReturn(maxBlur)
-        `when`(blurUtils.maxBlurRadius).thenReturn(maxBlur)
+        `when`(blurUtils.maxBlurRadius).thenReturn(maxBlur.toFloat())
+        `when`(blurUtils.maxBlurRadius).thenReturn(maxBlur.toFloat())
 
         notificationShadeDepthController =
             NotificationShadeDepthController(
@@ -322,8 +322,8 @@ class NotificationShadeDepthControllerTest : SysuiTestCase() {
         `when`(blurUtils.ratioOfBlurRadius(anyFloat())).then { answer ->
             answer.arguments[0] as Float / maxBlur.toFloat()
         }
-        `when`(blurUtils.maxBlurRadius).thenReturn(maxBlur)
-        `when`(blurUtils.maxBlurRadius).thenReturn(maxBlur)
+        `when`(blurUtils.maxBlurRadius).thenReturn(maxBlur.toFloat())
+        `when`(blurUtils.maxBlurRadius).thenReturn(maxBlur.toFloat())
 
         notificationShadeDepthController.transitionToFullShadeProgress = 1f
         notificationShadeDepthController.updateBlurCallback.doFrame(0)
